+++ wicket/src/main/java/org/apache/wicket/markup/html/form/Check.java
@@ -60,7 +60,7 @@ public class Check<T> extends LabeledWebMarkupContainer
 	 * page-scoped uuid of this check. this property must not be accessed directly, instead
 	 * {@link #getValue()} must be used
 	 */
-	private short uuid = -1;
+	private int uuid = -1;
 
 	private final CheckGroup<T> group;
